added compilations dir of ids/prefix counters generated (srav1/2,ccle); improved prefix counters file handling

+++ sample_ids/assign_compilation_ids.py
@@ -1,5 +1,6 @@
 #!/bin/env python2
 import sys
+import os
 import shutil
 import argparse
 from collections import Counter
@@ -8,7 +9,7 @@ import buildIDs
 COMPILATION_CODE_BIT_LENGTH=8
 
 def main(args):
-    #currently only converts a SRA run accession (e.g. SRR1234567) or a TCGA GDC File ID (UUID) (e.g. bef8b87c-4473-441a-a840-0bd8f2f1bc9e)
+    #currently only converts a SRA run accession (e.g. SRR1234567) or a TCGA/CCLE GDC File ID (UUID) (e.g. bef8b87c-4473-441a-a840-0bd8f2f1bc9e)
     counter = -1
     compilation_code = ""
     if args.compilation_code != -1:
@@ -19,6 +20,8 @@ def main(args):
             for line in fin:
                 (group, count) = line.rstrip().split('\t')
                 prefix_counters[str(group)] = int(count)
+    else:
+        args.group_counters_file = "%s.group_counters" % args.accessions_file
     fin = open(args.accessions_file, "rb")
     for line in fin:
         if counter == -1 and not args.no_header:
@@ -38,7 +41,8 @@ def main(args):
         #sys.stdout.write(str(int(prefix_ctr_str,2))+"\t"+acc+\t"+"\n")
         sys.stdout.write(fields[0]+"\t"+acc+"\t"+str(int(prefix_ctr_str,2))+"\n")
         counter+=1
-    shutil.move(args.group_counters_file, "%s.old" % (args.group_counters_file))
+    if os.path.exists(args.group_counters_file):
+        shutil.move(args.group_counters_file, "%s.old" % (args.group_counters_file))
     with open(args.group_counters_file, "w") as fout:
         [fout.write("%s\t%s\n" % (x, str(prefix_counters[x]))) for x in prefix_counters.keys()]
 
@@ -47,7 +51,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Assigns IDs for a specified compilation')
 
     parser.add_argument('--accessions-file', metavar='/path/to/file_with_accessions', type=str, default=None, help='path to file with list of accessions to assign IDs to')
-    parser.add_argument('--group-counters-file', metavar='/path/to/file_with_current_group_id_counters', type=str, default=None, help='path to file with the list of groups (low order digits) and their last used id counts')
+    parser.add_argument('--group-counters-file', metavar='/path/to/file_with_current_group_id_counters', type=str, default=None, help='path to file with the list of groups (low order digits) and their last used id counts; this code does not include the compilation prefix, so it is compilation specific')
     parser.add_argument('--acc-col', metavar='integer column #', type=int, default=0, help='column number in --accessions-file where accession is (defaults to 0)')
     parser.add_argument('--compilation-code', metavar='8-bit integer code per compilation', type=int, default=-1, help='used as a prefix to the sample ID to differentiate samples from different compilations when results are merged across compilations; 15 (1111) is reserved for local (non-public) compilations.  Default is not to include the compilation ID in the ID generation.')
     parser.add_argument('--no-header', action='store_const', const=True, default=False, help='if --accessions-file doesn\'t have a header')
